+++ HISTORY.rst
@@ -11,7 +11,7 @@ v3.0.33.dev (TBD)
   or :meth:`upload_filmstrip_thumbnail` before calling out to the server.
 - Multiple enhancements and bugfixes to Mockgun:
    - Added support for filters using the `filter_operator` syntax.
-   - `is` and `is_not` comparison with a field set to `None` are now handled.
+   - `is` and `is_not` comparisons with an entity or multi-entity field set to `None` are now handled.
    - Fixed contains operator on strings.
    - Fixed entity link comparisons with `None`.
    - Added support for dotted notation in find parameters (e.g. `users.HumanUser.login`) on both entity and multi-entity fields.